Refactor homePage to use custom hook
- useFilterCountries performs filtering logic
- simplifies funciton body of HomePage
